% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData.R
\name{add_grid}
\alias{add_grid}
\title{Add grid to LPJmLData object}
\usage{
add_grid(x, ...)
}
\arguments{
\item{...}{arguments passed to \code{\link{read_io}} if no grid file and or meta
file in corresponding output directory available.}
}
\value{
\code{\link{LPJmLData}} object in selected format
}
\description{
Function to add a grid to an \code{\link{LPJmLData}}. The function acts
as a \code{\link{read_io}} function for the grid file and adds it as an
\code{LPJmLData} object itself to the the main object as an attribute \verb{$grid}
}
\examples{
\dontrun{

vegc <- read_io(filename = "./vegc.bin.json")

}

}
