% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{aggregateLPJmLdata2Country}
\alias{aggregateLPJmLdata2Country}
\title{Aggregate LPJmL to country level}
\usage{
aggregateLPJmLdata2Country(input, cowList, aggMethod = "sum")
}
\arguments{
\item{input}{lpjml array to be aggregated c(ncells)}

\item{cowList}{country-list giving the country code for each LPJmL cell dim=c(ncells)}

\item{aggMethod}{method to aggregate data to country level ("sum" - e.g. for precipitation - or "mean" - e.g. for temperatures)}
}
\value{
country-array
}
\description{
Aggregates all data in lpjml format to country level, using the aggregation Method provided (e.g. sum, mean)
}
\examples{
aggregateLPJmLdata2Country(input=surface_temperature,cowList=cowList,aggMethod="mean")

}
