% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLMetaData_as.R
\name{as_header}
\alias{as_header}
\title{Coerce LPJmLMetaData to a LPJmL header object}
\usage{
as_header(x, silent = TRUE)
}
\arguments{
\item{x}{\link{LPJmLMetaData} object}

\item{silent}{Logical. Whether to suppress notifications from header
conversion/initialization.}
}
\value{
a LPJmL header object, more information at \code{\link{create_header}}
}
\description{
Function to coerce (convert) a \code{LPJmLMetaData} object into a
LPJmL header object, more information at \code{\link{create_header}}.
}
\examples{
\dontrun{

vegc_meta <- read_meta(filename = "./vegc.bin.json")

# returns one dimensional array with timeseries for cells `27410:27415`
as_header(vegc_meta)
# $name
# [1] "LPJDUMMY"
#
# $header
#      version        order    firstyear        nyear    firstcell
#          4.0          4.0       1901.0        200.0          0.0
#       nbands cellsize_lon       scalar cellsize_lat     datatype
#          1.0          0.5          1.0          0.5          3.0
#     timestep
#          1.0
#
# $endian
# [1] "little"

}

}
