% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_as.R
\name{as_raster}
\alias{as_raster}
\title{Coerce LPJmLData to a raster object}
\usage{
as_raster(x, subset = NULL, aggregate = NULL, ...)
}
\arguments{
\item{x}{\link{LPJmLData} object}

\item{subset}{list of array dimension(s) as name/key and
corresponding subset vector as value, e.g.
\verb{list(cell = c(27411:27415)}, more information at
\code{\link{subset.LPJmLData}}.}

\item{aggregate}{list of array dimension(s) as name/key and
corresponding aggregation function as value, e.g.
\code{list(band = sum)}.}

\item{...}{arguments forwarded to the aggregate function(s), e.g.
\code{na.rm = TRUE}.}
}
\value{
a \link[raster]{raster} or \link[raster]{brick} with grid based on
internal \verb{$grid} attribute (\code{\link{LPJmLData}} of \code{"./grid.*"}) and corresponding
data for each grid cell. If multiple bands or time dimensions exist, a
\link[raster]{brick} is created. Further meta information such as the
lon/lat resolution are extracted from \verb{$meta}.
}
\description{
Function to coerce (convert) a \code{\link{LPJmLData}} object into a
\link[raster]{raster} or \link[raster]{brick} object, that opens the space
for any GIS based raster operations. Read more
\href{https://rspatial.github.io/raster/reference/raster-package.html}{here}. The
successor package of raster is \href{https://rspatial.org/}{terra}.
}
\details{
for coercion to a \link[raster]{raster} or \link[raster]{brick},
the \verb{$grid} attribute is required. When using \code{file_type = "meta"}, grid data
is read automatically via \code{\link{add_grid}}. If \code{file_type = "clm"}
or \code{file_type = "raw"} is used, \code{add_grid} has to be performed beforehand.
}
\examples{
\dontrun{

vegc <- read_io(filename = "./vegc.bin.json")

# returns a raster brick for all data
as_raster(vegc)
# class      : RasterBrick
# dimensions : 280, 720, 201600, 200  (nrow, ncol, ncell, nlayers)
# resolution : 0.5, 0.5  (x, y)
# extent     : -180, 180, -56, 84  (xmin, xmax, ymin, ymax)
# crs        : +proj=longlat +datum=WGS84 +no_defs
# source     : memory
# names      : X1901.12.31, X1902.12.31, X1903.12.31, X1904.12.31, ...
# min values :           0,           0,           0,           0, ...
# max values :    28680.72,    28662.49,    28640.29,    28634.03, ...

}

}
