% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_as.R
\name{as_tibble}
\alias{as_tibble}
\title{Coerce LPJmLData to an tibble}
\usage{
as_tibble(x, ...)
}
\arguments{
\item{...}{arguments forwarded to the aggregate function(s), e.g.
\code{na.rm = TRUE}.}

\item{subset}{list of array dimension(s) as name/key and
corresponding subset vector as value, e.g.
\verb{list(cell = c(27411:27415)}, more information at
\link[lpjmlkit]{subset}.}

\item{aggregate}{list of array dimension(s) as name/key and
corresponding aggregation function as value, e.g.
\code{list(band = sum)}.}

\item{value_name}{rename value column in  returned \code{tibble}. Defaults to
\code{"value"}.}
}
\value{
a \verb{\link[tibble]\{tibble\}} with columns corresponding to dimension
naming of \code{LPJmLData$data} array and values in value column.
}
\description{
Function to coerce (convert) a \code{LPJmLData} object into a
\link[tibble]{tibble} (modern \verb{\link[base]\{data.frame\}}, read more
\href{https://r4ds.had.co.nz/tibbles.html}{here})
}
\examples{
\dontrun{

vegc <- read_io(filename = "./vegc.bin.json")

# returns one dimensional array with timeseries for cells `27410:27415`
as_tibble(vegc)
#   cell  time       band    value
#   <fct> <fct>      <fct>   <dbl>
# 1 0     1901-12-31 1     13627.
# 2 1     1901-12-31 1       120.
# 3 2     1901-12-31 1       133.
# 4 3     1901-12-31 1        97.4
# 5 4     1901-12-31 1        75.9
# 6 5     1901-12-31 1        11.1

}

}
