% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_array.R
\name{asub}
\alias{asub}
\alias{asub<-}
\title{Subset a named array}
\usage{
asub(x, ..., drop = TRUE)

asub(x, ...) <- value
}
\arguments{
\item{x}{An array with named dimensions}

\item{...}{Provide dimension names to be used to subset an \link[base]{array}
in combination with indices vectors, e.g. \code{cell = c(27411:27416)}, or
\code{band = -c(14:16, 19:32)} or subset using a "character" vector like
\code{band = c("rainfed rice","rainfed maize")}.}

\item{drop}{Logical. If TRUE (default), dimensions are dropped when dimension
has length == 1, else dimension is kept.}

\item{value}{Array/vector of the same dimension as referred to by dimension
and subset vector (\code{...}).}
}
\value{
Array or vector (if \code{drop=TRUE} and one only one dimension is left)
}
\description{
Subset an array with the supplied dimnames and - if defined - replace values.
}
\section{Functions}{
\itemize{
\item \code{asub(x, ...) <- value}: replace an array subset

}}
\examples{
my_array <- array(1,
                  dim=c(cell = 67, month=12, band = 3),
                  dimnames = list(cell = 0:66,
                                month = 1:12,
                                band = c("band1", "band2", "band3")))
my_subset <- asub(my_array,
                  band = c("band1", "band3"))
dimnames(my_subset)[3]
# $ band
#   [1] "band1"
#   [2] "band3"

# replace subset
asub(my_subset, band = c("band1")) <- 0

}
