% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{autoReadInput}
\alias{autoReadInput}
\title{Automatically reads LPJmL input including header}
\usage{
autoReadInput(
  inFile,
  getyearstart = -1,
  getyearstop = -1,
  manu = F,
  msize = 4,
  mheadersize = 43
)
}
\arguments{
\item{inFile}{name of file to read}

\item{getyearstart}{first year to be read (if not specified will default to startyear from header)}

\item{getyearstop}{final year to be read (if not specified will default to stopyear from header)}

\item{manu}{switch to allow for manual override of size (default False)}

\item{msize}{manual size, only applied if manu==TRUE (default 4)}

\item{mheadersize}{manual header size, only applied if manu==TRUE (default 43)}
}
\value{
data array
}
\description{
most header variables will be automatically read
for some combinations this won't work, thus
a manual override switch can be used to set size
}
\examples{
\dontrun{
autoReadInput(inFile,getyearstart=-1,getyearstop=-1,manu=T,msize=2,mheadersize=0)
}

}
