% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cellarea.R
\name{calc_cellarea}
\alias{calc_cellarea}
\title{Calculate cell area of LPJmL cells}
\usage{
calc_cellarea(
  x,
  res_lon = 0.5,
  res_lat = res_lon,
  earth_radius = 6371000.785,
  unit = "m2"
)
}
\arguments{
\item{x}{\code{LPJmLData} object with \verb{$grid} attribute, a LPJmLData object of
variable \code{"grid"} (\code{"LPJ_GRID"}) or vector of cell-center latitude
coordinates in degrees.}

\item{res_lon}{Grid resolution in longitude direction in degrees
(default: 0.5). If x LPJmLData` object it will be ignored if different.}

\item{res_lat}{Grid resolution in latitude direction in degrees (default:
same as res_lon). If x LPJmLData` object it will be ignored if different.}

\item{earth_radius}{Radius of sphere (in \eqn{m}) used to calculate cell
areas.}

\item{unit}{Character string, defaults to \code{"m2"}. Further available options
are \code{"ha"} or \code{"km2"}}
}
\value{
Either if x is an LPJmLData object a vector or array matching the
space dimension(s) of this object or if x is a latitude vector a vector of
cell areas in \eqn{m^2} corresponding to cells in \code{lat} is returned.
}
\description{
Calculate cell area of LPJmL cells based on a LPJmLData object
or cell coordinates and grid resolution.
Uses a spherical representation of Earth.
}
\examples{
grid <- matrix(
  data = c(-179.75, 89.75, -0.25, 0.25, 0.25, -0.25, 179.75, -89.75),
  ncol = 2,
  byrow = TRUE,
  dimnames = list(NULL, c("lon", "lat"))
)
gridarea <- calc_cellarea(grid[,"lat"])

}
