% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEFRs.R
\name{calcEFRs}
\alias{calcEFRs}
\title{Calculate environmental flow requirements (EFRs)}
\usage{
calcEFRs(discharge, method = "VMF")
}
\arguments{
\item{discharge}{30 years monthly discharge array dim(discharge)=c(ncells,12,30)}

\item{method}{EFR method to be used ("VMF","Q90Q50","PBpaper"), "PBpaper" is the modified version of VMF used for the Steffen et al. 2015 PB paper}
}
\value{
EFRs as volumes like in discharge input, dim(EFRs)=c(ncells,12)
}
\description{
Calculate environmental flow requirements (EFRs), based on 30 years of monthly discharge input
}
\examples{
\dontrun{
calcEFRs(discharge=mdischarge_preindustrial,method="VMF")
}

}
