% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_area.R
\name{cell_area}
\alias{cell_area}
\title{Calculate cell area of LPJmL cells}
\usage{
cell_area(lat, res_lon = 0.5, res_lat = res.lon, earth_radius = 6371000.785)
}
\arguments{
\item{lat}{Vector of cell-center latitude coordinates in degrees}

\item{res_lon}{Grid resolution in longitude direction in degrees (default: 0.5)}

\item{res_lat}{Grid resolution in latitude direction in degrees (default: same as res_lon)}

\item{earth_radius}{Radius of sphere (in \eqn{m}) used to calculate cell areas}
}
\value{
Vector of cell areas in \eqn{m²} corresponding to cells in \code{lat}
}
\description{
Calculate cell area of LPJmL cells based on
cell coordinates and grid resolution. Uses a spherical representation of Earth.
}
\examples{
grid <- matrix(data=c(-179.75, 89.75, -0.25, 0.25, 0.25, -0.25, 179.75, -89.75),
               ncol=2,
               byrow=TRUE,
               dimnames=list(NULL, c("lon", "lat"))
              )
gridarea <- cell_area(grid[,"lat"])

}
