% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellarea.R, R/process.R
\docType{data}
\name{cellarea}
\alias{cellarea}
\title{LPJmL cellarea}
\format{
An array of dim (67420), with the area in m2 for each LPJmL grid cell.
}
\source{
contact Fabian Stenzel
}
\usage{
data(cellarea)

cellarea(lat, res.lon = 0.5, res.lat = res.lon)
}
\arguments{
\item{lat}{vector of latutidinal grid-midpoints}

\item{res.lon}{resolution (degree) in longitude dimension (default 0.5)}

\item{res.lat}{resolution (degree) in latitude dimension (default res.lon)}
}
\value{
vector of cellsizes in m2
}
\description{
This file contains data about LPJmL 
cellsizes in m2

this function computes gridcell areas in m2 based on latitude
coordinate and angular cell size (resolution), assuming the earth
to be a sphere
lat can be a vector, res.lon and res.lat be single
values.
}
\examples{
data(cellarea)
\dontrun{
cellarea(header=header)
}

}
\references{
See LPJmL gitlab wiki
(\href{https://gitlab.pik-potsdam.de/lpjml/LPJmL_internal/wikis/Output}{LPJmL wiki Output})
}
\keyword{datasets}
