% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_names.R
\name{create_time_names}
\alias{create_time_names}
\title{Create a character vector of dates.}
\usage{
create_time_names(nstep = 365, years = 2000)
}
\arguments{
\item{nstep}{An integer value defining the time step of the output file.
Valid values are 1 (yearly), 12 (monthly), 365 (daily).}

\item{years}{An integer vector of (sequential or non-sequential) years.}
}
\value{
A character vector of dates in the format "YYYY-MM-DD".
}
\description{
Create a vector of dates to be used as
e.g. names of a time dimension. Dates are written as.characters, not as.Date,
in order to avoid issues with leap years. By default, for annual/monthly outputs,
it returns the last day of the year/month, as these are the days when LPJmL
writes annual/montly data outputs.
}
\seealso{
strptime, as.Date, format
}
