% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{daily2monthly}
\alias{daily2monthly}
\title{Aggregate from daily values to monthly values}
\usage{
daily2monthly(daily, method = "sum")
}
\arguments{
\item{daily}{array with daily values dim(daily)=c(365)}

\item{method}{function for aggregation: sum or mean}
}
\value{
monthly array
}
\description{
Function to aggregate from daily values to monthly values. Can also be applied for several year arrays:
apply(X = daily_discharge,MARGIN = c(1,3),FUN = daily2monthly,method="sum")
with dim(daily_discharge)=c(67420,365,30)
}
\examples{
\dontrun{
daily2monthly(daily=ddischarge,method="sum")
}

}
