% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{dimnames.LPJmLData}
\alias{dimnames.LPJmLData}
\title{Dimnames of a LPJmLData's data array
Function to get the dimensions (list) of the array of a
\code{LPJmLData} object.}
\usage{
\method{dimnames}{LPJmLData}(x, ...)
}
\arguments{
\item{x}{\link{LPJmLData} object}

\item{...}{further arguments to be passed on to \link[base]{dimnames}}
}
\description{
Dimnames of a LPJmLData's data array
Function to get the dimensions (list) of the array of a
\code{LPJmLData} object.
}
