% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{dn}
\alias{dn}
\title{Shortcut dimnames (type-corrected) of an LPJmLData data array}
\usage{
dn(x)
}
\arguments{
\item{x}{\link{LPJmLData} object}
}
\description{
Function to get the dimnames (list) of the data array of an LPJmLData object
as numeric where vector where possible.
}
