% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenGrid2lpjGrid.R
\name{evenGrid2lpjGrid}
\alias{evenGrid2lpjGrid}
\title{Transform even grid to lpjml grid}
\usage{
evenGrid2lpjGrid(gridIn, lonIn, latIn, ncells = 67420, res = 0.5)
}
\arguments{
\item{gridIn}{even grid to be transformed (e.g. c(720,360) for 0.5 deg)}

\item{lonIn}{longitude vector of input grid}

\item{latIn}{latitude vector of input grid}

\item{ncells}{number of lpj cells (default is 67420 for 30min res), 2298847 for 5min res}

\item{res}{resolution (default 0.5)}
}
\value{
outlist transformed data in lpjml format
}
\description{
Transform even grid (e.g. from a netcdf) to lpjml grid
}
\examples{
evenGrid2lpjGrid(gridIn=climateCategories,lonIn=seq(-179.75,179.75,0.5),latIn=seq(-89.75,89.75,0.5))

}
