% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{flowSeason}
\alias{flowSeason}
\title{Return flow season}
\usage{
flowSeason(discharge, method = "VMF")
}
\arguments{
\item{discharge}{30 years monthly discharge array dim(discharge)=c(ncells,12,30)}

\item{method}{EFR method to be used ("VMF","Q90Q50","PBpaper"), "PBpaper" is the modified version of VMF used for the Steffen et al. 2015 PB paper}
}
\value{
flow season array:  "l" (low flow), "i" (intermediate - if applicable), or "h" (high flow)
}
\description{
Return flow season based on the chosen EFR method. "l" (low flow), "i" (intermediate - if applicable), or "h" (high flow)
}
\examples{
\dontrun{
flowSeason(discharge=mdischarge_preindustrial,method="VMF")
}

}
