% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadheader.R
\name{freadheader}
\alias{freadheader}
\title{Get header of LPJmL input file}
\usage{
freadheader(f.in)
}
\arguments{
\item{f.in}{file connection to read the header from}
}
\value{
data_frame of header data
h$v : header version,
h$o : order,
h$fy : firstyear,
h$ny : nyear,
h$fc : firstcell,
h$nc : ncells,
h$nb : nbands,
h$r : resolution,
h$b : scaling
}
\description{
Reads the header of a binary CLM(2) LPJmL input file and returns the content as data_frame
}
\examples{
freadheader(system.file("extdata", "grid.bin", package = "lpjmliotools", mustWork = TRUE))

}
