% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwriteheader.R
\name{fwriteheader}
\alias{fwriteheader}
\title{Write header of LPJmL input file}
\usage{
fwriteheader(
  file.out,
  headername,
  version = 2,
  firstyear,
  lastyear,
  bands,
  scalar,
  ncells = 67420
)
}
\arguments{
\item{file.out}{file connection to the file to append the header to}

\item{headername}{character string containing the name of the header}

\item{version}{CLM Version (default is 2)}

\item{firstyear}{first year of data set}

\item{lastyear}{last year covered in the data set}

\item{bands}{number of cft-bands (32 or 64 for standard input)}

\item{scalar}{for cft fracs set to 0.001}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}
}
\value{
None
}
\description{
Writes the header of a binary CLM(2) LPJmL input file
into an open file connection
}
\examples{
\dontrun{
fwriteheader(file.out=cftfracs_file,headername="LPJLUSE",version=2, firstyear = 2006,lastyear = 2100, bands=64, scalar=0.001)
}

}
