% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_size.R
\name{header_size}
\alias{header_size}
\title{Determine size of file header}
\usage{
header_size(header)
}
\arguments{
\item{header}{Header list object as returned by read_header() or new_header()}
}
\value{
Integer value giving the size of the header in bytes. This can be used when seeking in the file.
}
\description{
Returns the size in bytes of an LPJmL file header based on
a header list object read by read_header() or generated by new_header()
}
\examples{
\dontrun{
header <- read_header("filename.clm")
size <- header_size(header)
# open file for reading
fp <- file("filename.clm", "rb")
# skip over file header
seek(fp, size)
# add code to read data from file
}
}
\seealso{
\code{\link[=read_header]{read_header()}} for reading a header from an LPJmL file, \code{\link[=new_header]{new_header()}} fore creating a header from scratch.
}
