% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headerSize.R
\name{headerSize}
\alias{headerSize}
\title{Determine size of file header}
\usage{
headerSize(header)
}
\arguments{
\item{header}{Header list object as returned by readHeader() or newHeader()}
}
\value{
Integer value giving the size of the header in bytes. This can be used when seeking in the file.
}
\description{
Returns the size in bytes of an LPJmL file header based on
a header list object read by readHeader() or generated by newHeader()
}
\examples{
\dontrun{
header <- readHeader("filename.clm")
size <- headerSize(header)
# open file for reading
fp <- file("filename.clm", "rb")
# skip over file header
seek(fp, size)
# add code to read data from file
}
}
\seealso{
\code{\link[=readHeader]{readHeader()}} for reading a header from an LPJmL file, \code{\link[=newHeader]{newHeader()}} fore creating a header from scratch.
}
