% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat.R
\docType{data}
\name{lat}
\alias{lat}
\title{LPJmL latitude}
\format{
An array of dim (67420), with the latitude for each LPJmL grid cell.
}
\source{
contact Fabian Stenzel
}
\usage{
data(lat)
}
\description{
This file contains the latitude of the midpoint for each grid cell
}
\examples{
data(lat)
}
\references{
See LPJmL gitlab wiki
(\href{https://gitlab.pik-potsdam.de/lpjml/LPJmL_internal/wikis/Output}{LPJmL wiki Output})
}
\keyword{datasets}
