% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData.R
\name{length.LPJmLData}
\alias{length.LPJmLData}
\title{Get length of LPJmLData's data array
Function to get the length of the array of a LPJmLData object}
\usage{
\method{length}{LPJmLData}(x)
}
\arguments{
\item{...}{further arguments to be passed on to \link[base]{length}}
}
\description{
Get length of LPJmLData's data array
Function to get the length of the array of a LPJmLData object
}
