% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lon.R
\docType{data}
\name{lon}
\alias{lon}
\title{LPJmL longitude}
\format{
An array of dim (67420), with the longitude for each LPJmL grid cell.
}
\source{
contact Fabian Stenzel
}
\usage{
data(lon)
}
\description{
This file contains the latitude of the midpoint for each grid cell
}
\examples{
data(lon)
}
\references{
See LPJmL gitlab wiki
(\href{https://gitlab.pik-potsdam.de/lpjml/LPJmL_internal/wikis/Output}{LPJmL wiki Output})
}
\keyword{datasets}
