% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{lpj2raster}
\alias{lpj2raster}
\title{Create a raster object from an LPJmL data array}
\usage{
lpj2raster(lpjarray, lat, lon)
}
\arguments{
\item{lpjarray}{lpjml array filled with data dim(lpjmlarray)==c(ncells). ncells==67420 for 30min resolution, or ncells==2298847 for 5min resolution}

\item{lat}{latitude vector of lpj array}

\item{lon}{longitude vector of lpj array}
}
\value{
ra raster object filled with the lpjml data on land
}
\description{
Create a raster object from an LPJmL data array and the corresponding lat/lon arrays
}
\examples{
\dontrun{
lpj2raster(lpjarray=discharge2000,lat=lat,lon=lon)
}

}
