% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{LPJ_datatype}
\alias{LPJ_datatype}
\title{Get R-specific datatype and size from LPJmL header CLM3}
\usage{
LPJ_datatype(header)
}
\arguments{
\item{header}{header list}
}
\value{
list(datatype,size)
}
\description{
This function only works properly with headers of CLM version3 !
it can be used within readBin()
given an LPJ file header (as returned by readheader()),
it returns a list with 2 items:
- datatype (to be used as the "what" parameter in readBin()
- size (to be used as the "size" parameter in readBin()
datatypes as currently implemented:
0: character (size 1)
1: short (integer size 2)
2: integer (integer size 4)
3: float (numeric size 4)
4: double (numeric size 8)
}
\examples{
\dontrun{
LPJ_datatype(header=header)
}

}
