% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjCOW.R
\name{lpjCOW}
\alias{lpjCOW}
\title{Get country code list}
\usage{
lpjCOW(fileName, ncells = 67420)
}
\arguments{
\item{fileName}{character string containing the file to read the cow from ()}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}
}
\value{
array of 2 IDs (countrycode,regioncode) for each of the 67420 gridcells
}
\description{
Reads and returns the list of country codes contained in
}
\examples{
lpjCOW(system.file("extdata", "cow_mg_2006.bin", package = "lpjmliotools", mustWork = TRUE))

}
