% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjDatatype.R
\name{lpjDatatype}
\alias{lpjDatatype}
\title{Data type and size}
\usage{
lpjDatatype(header)
}
\arguments{
\item{header}{Header list object as returned by readHeader() or newHeader()}
}
\value{
The function returns a list object with three components:
\itemize{
\item type: R data type; can be used with 'what' parameter of readBin()
\item size: size of data type; can be used with 'size' parameter of readBin()
\item signed: whether or not the data type is signed, only relevant for integers with size 1 or 2; can be used with 'signed' parameter of readBin()
}
}
\description{
Provide information on the data type used in an LPJmL file
based on the 'datatype' attribute included in the file header.
}
\examples{
\dontrun{
# read file header
header <- readHeader("filename.clm")
# open file for reading
fp <- file("filename.clm", "rb")
# skip over file header
seek(fp, headerSize(header))
# read in file data
fileData <- readBin(fp, what=lpjDatatype(header)$type, size=lpjDatatype(header)$size, signed=lpjDatatype(header)$signed, n=header[["header"]]["ncell"]*header[["header"]]["nbands"]*header[["header"]]["nyear"], endian=header[["endian"]])
# close file
close(fp)
}

}
\seealso{
\code{\link[=readHeader]{readHeader()}} for reading headers from LPJmL files, \code{\link[=newHeader]{newHeader()}} for creating headers from scratch
}
