% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjmlkit-package.R
\docType{package}
\name{lpjmlkit-package}
\alias{lpjmlkit}
\alias{lpjmlkit-package}
\title{lpjmlkit: Toolkit for Basic LPJmL Handling}
\description{
A collection of basic functions to facilitate the work with the Dynamic Global Vegetation Model (DGVM) Lund-Potsdam-Jena managed Land (LPJmL) hosted at the Potsdam Institute for Climate Impact Research (PIK). It provides functions for performing LPJmL simulations, as well as reading, processing and writing model-related data such as inputs and outputs or configuration files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PIK-LPJmL/lpjmlkit}
  \item \doi{10.5281/zenodo.7773134}
  \item Report bugs at \url{https://github.com/PIK-LPJmL/lpjmlkit/issues}
}

}
\author{
\strong{Maintainer}: Jannes Breier \email{jannesbr@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9055-6904}{ORCID})

Authors:
\itemize{
  \item Jannes Breier \email{jannesbr@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9055-6904}{ORCID})
  \item Sebastian Ostberg \email{ostberg@pik-potsdam.de} (\href{https://orcid.org/0000-0002-2368-7015}{ORCID})
  \item Stephen Björn Wirth \email{wirth@pik-potsdam.de} (\href{https://orcid.org/0000-0003-3090-3318}{ORCID})
  \item Sara Minoli \email{minoli@pik-potsdam.de} (\href{https://orcid.org/0000-0001-7920-3107}{ORCID})
  \item Fabian Stenzel \email{stenzel@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5109-0048}{ORCID})
  \item David Hötten \email{davidho@pik-potsdam.de}
  \item Christoph Müller \email{cmueller@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9491-3550}{ORCID})
}

}
