% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lpjml.R
\name{make_lpjml}
\alias{make_lpjml}
\title{Compile LPJmL model}
\usage{
make_lpjml(
  model_path = ".",
  make_fast = TRUE,
  make_clean = FALSE,
  throw_error = TRUE,
  debug = FALSE
)
}
\arguments{
\item{model_path}{Character string providing the path to LPJmL
(equal to \code{LPJROOT} environment variable)}

\item{make_fast}{Logical. If set to \code{TRUE}, calls "make -j16 all" instead of
"make all" for compilation in parallel mode on 16 CPUs. Defaults to \code{TRUE}.}

\item{make_clean}{Logical. If set to \code{TRUE}, calls "make clean" first to
remove previous installation. Defaults to \code{FALSE}.}

\item{throw_error}{Logical. Whether to throw an error if sub-process has
non-zero exit status, hence if compilation fails. Defaults to \code{TRUE}.}

\item{debug}{Logical. Whether to compile LPJmL with \code{-debug} flag. Defaults to
\code{FALSE}.}
}
\value{
A list with process status, see \link[processx]{run}.
}
\description{
Compiles the LPJmL source code and creates an executable by executing
"make all" on the operating system shell.
}
\examples{
\dontrun{
model_path <- "./LPJmL_internal"
make_lpjml(model_path = model_path)
}

}
