% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdfDaily2lpjarray.R
\name{netcdfDaily2lpjarray}
\alias{netcdfDaily2lpjarray}
\title{Reads daily netcdf and returns it as LPJ-array}
\usage{
netcdfDaily2lpjarray(ncInFile, var, years, lon, lat)
}
\arguments{
\item{ncInFile}{netcdf file name}

\item{var}{netcdf variable name}

\item{years}{total years stored in netcdf file}

\item{lon}{array with longitude coordinates of the centers of the LPJmL grid}

\item{lat}{array with latitude coordinates of the centers of the LPJmL grid}
}
\value{
lpjml array with data from netcdf, dim=c(ncells,days,years)
}
\description{
Reads a daily netcdf and returns the values at the LPJ-grid locations
the netcdf resolution needs to correspond to the LPJmL grid resolution
(i.e. the LPJ cells need to have the same center coordinates than the netcdf)
}
\examples{
\dontrun{
netcdfDaily2lpjarray(ncInFile=ddischarge_1901-2005.nc4,var="dis",years=105,lon=lon,lat=lat)
}

}
