% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{netcdfMonthly2lpjarray}
\alias{netcdfMonthly2lpjarray}
\title{Reads monthly netcdf and returns it as LPJ-array}
\usage{
netcdfMonthly2lpjarray(ncInFile, var, lon, lat)
}
\arguments{
\item{ncInFile}{netcdf file name}

\item{var}{netcdf variable name}

\item{lon}{array with longitude coordinates of the centers of the LPJmL grid}

\item{lat}{array with latitude coordinates of the centers of the LPJmL grid}
}
\value{
lpjml array with data from netcdf, dim=c(ncells,months,years)
}
\description{
Reads a monthly netcdf and returns the values at the LPJ-grid locations
the netcdf resolution needs to correspond to the LPJmL grid resolution
(i.e. the LPJ cells need to have the same center coordinates than the netcdf)
}
\examples{
\dontrun{
netcdfMonthly2lpjarray(ncInFile=mdischarge_1901-2005.nc4,var="dis",years=105,lon=lon,lat=lat)
}

}
