% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{new_header}
\alias{new_header}
\title{Creates a CLM header list-object for LPJmL}
\usage{
new_header(
  name = "LPJGRID",
  version = 3,
  order = 1,
  firstyear = 1901,
  nyear,
  firstcell = 0,
  ncell,
  nbands,
  cellsize_lon = 0.5,
  scalar = 1,
  cellsize_lat = cellsize_lon,
  datatype = 1,
  endian = .Platform$endian
)
}
\arguments{
\item{name}{header name (default "LPJGRID")}

\item{version}{header CLM version (default 2)}

\item{order}{order of data items , either CELLYEAR (1) or YEARCELL (2) (default 1)}

\item{firstyear}{first year for data (default 1901)}

\item{nyear}{number of years contained in file}

\item{firstcell}{index of first data item (default 0)}

\item{ncell}{number of data items per year}

\item{nbands}{number of data elements per cell}

\item{cellsize_lon}{longitude cellsize in deg (default 0.5)}

\item{scalar}{conversion factor (default 1)}

\item{cellsize_lat}{latitude cellsize in deg (default 0.5)}

\item{endian}{endianness (default copied from local platform -- .Platform$endian)}

\item{datatyp}{data type in file (default 1)}
}
\value{
list-object with header content
}
\description{
Creates a header from scratch which can be saved to a file using
function writeheader()
Parameters without defaults that you always need to provide:
nyear, ncell, nbands
}
\examples{
\dontrun{
new_header(name="LPJGRID", version=2, order=1, firstyear=1901, nyear=1, firstcell=0, ncell=67420, nbands=1, cellsize_lon=0.5, scalar=1, cellsize_lat=cellsize_lon, datatype=1, endian=.Platform$endian)
}

}
