% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pie2}
\alias{pie2}
\title{Copy of the pie function with labels further away from pie}
\usage{
pie2(
  x,
  labels = names(x),
  edges = 200,
  radius = 0.8,
  clockwise = FALSE,
  init.angle = if (clockwise) 90 else 0,
  density = NULL,
  angle = 45,
  col = NULL,
  border = NULL,
  lty = NULL,
  main = NULL,
  ...
)
}
\description{
Copy of the pie function with labels further away from pie
}
