% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_plot.R
\name{plot.LPJmLData}
\alias{plot.LPJmLData}
\title{Plot LPJmLData class}
\usage{
plot.LPJmLData(x, ...)
}
\arguments{
\item{...}{arguments forwarded to graphics::plot() and
raster::plot()}

\item{subset}{list of array dimension(s) as name/key and
corresponding subset vector as value, e.g.
\verb{list(cell = c(27411:27415)}, more information at
\link[lpjmlKit]{subset}.}

\item{aggregate}{list of array dimension(s) as name/key and
corresponding aggregation function as value, e.g.
\code{list(band = sum)}.}

\item{raster_extent}{if \code{aggregate} does not include space dim
\href{\link[raster]{extent}}{extent object}, or any object from
which an Extent object can be extracted}
}
\description{
Function to plot a time-series or raster map of a LPJmLData
object. This function is only supposed to give an overview of
the data and not to create publication ready graphics.
}
\examples{

}
