% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_relDiff}
\alias{plot_relDiff}
\title{Plot global relative differences}
\usage{
plot_relDiff(file, diff_data, min, max, negcol, poscol, title, legendtitle)
}
\arguments{
\item{file}{character string for location/file to save plot to}

\item{diff_data}{array with relative differences in LPJmL specific array c(67420)}

\item{min}{lower end of data range (between -1.5 and 0, with one digit allowed to the left of the decimal point); one additional break is added to display values lower than min}

\item{max}{upper end of data range (between 0 and 1.5, with one digit allowed to the left of the decimal point); one additional break is added to display values higher than max}

\item{negcol}{color palette name from brewer.pal (character string) for negative values}

\item{poscol}{color palette name from brewer.pal (character string) for poitive values}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}
}
\value{
None
}
\description{
Plot of relative differences derived from two LPJmL arrays
            classification into 10% classes
}
\examples{
\dontrun{
plot_relDiff(file="~/difference_in_soil_carbon_2005_2006.png",diff_data=diff_soilc_2005_2006, min=-0.5, max=0.7, negcol="Reds", poscol="Blues",
             title = "relative difference in soil carbon between 2005 and 2006",legendtitle="diff. in \%",eps=FALSE)
}

}
