% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBasin2screen}
\alias{plotBasin2screen}
\title{Plot basin to screen}
\usage{
plotBasin2screen(cell, endcellList, zoom = F, off = 20, ncells = 67420)
}
\arguments{
\item{cell}{index of the cell, whose basin should be plotted}

\item{endcellList}{lpjml array giving the index of the final drainage cell for each cell}

\item{zoom}{zoom in? (boolean) otherwise global plot}

\item{off}{offset amount in degrees to each side of the cell, only applied of zoom==T (defaults to 20)}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}
}
\value{
None
}
\description{
Plot all cells of a basin for given cell index to R internal screen
target cell is marked red, basin cells blue
}
\examples{
plotBasin2screen(cell=20000,endcellList=endcell,zoom=F,off=20)

}
