% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBEharvest}
\alias{plotBEharvest}
\title{Plot bioenergy harvest over time}
\usage{
plotBEharvest(file, beHarv, beHarv_cal, title, expFormat = "png")
}
\arguments{
\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{beHarvest}{array yearly beharvest data -- array c(years,type,clim,rf:irr)}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Plot bioenergy harvest over time
}
\examples{
\dontrun{
plotBEharvest(file="beHarvest.png",beHarvest=beHarvestYearly,
             title = "",eps=FALSE)
}

}
