% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCountryData}
\alias{plotCountryData}
\title{Plot global country data}
\usage{
plotCountryData(
  data,
  cowList,
  file,
  sty = "lin",
  title = "",
  legendtitle = "",
  ncells = 67420
)
}
\arguments{
\item{data}{country array to be plotted c(ncountries)}

\item{cowList}{country-list giving the country code for each LPJmL cell dim=c(ncells)}

\item{file}{character string for location/file to save plot to}

\item{sty}{array with breaks to be used}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}

\item{palette}{color palette string (defaults to "YlGnBu")}
}
\value{
None
}
\description{
Plot an array of country data, e.g. obtained from aggregateLPJmLdataCountry()
}
\examples{
plotCountryData(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep=""),
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)

}
