% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGlobal_pos}
\alias{plotGlobal_pos}
\title{Plot global LPJmL array with only positive values}
\usage{
plotGlobal_pos(data, file, title, pow2max, col, legendtitle, legYes, eps)
}
\arguments{
\item{data}{array with data to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{pow2max}{upper (positive) end of data range to plot (2^pow2max)}

\item{col}{color palette name from brewer.pal (character string)}

\item{legendtitle}{character string legend title}

\item{legYes}{show legend (boolean)}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Creates a PNG/eps with a plot of a global LPJmL array
   Data is plotted in range: c(0,2^pow2max)
   positive values are colored accoding to chosen palette,
   0-range is white,
}
\examples{
\dontrun{
plotGlobal_pos(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep=""),pow2max=15,col="GnBu",
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)
}
}
