% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGlobal_pos_neg}
\alias{plotGlobal_pos_neg}
\title{Plot global LPJmL array with positive and negative values}
\usage{
plotGlobal_pos_neg(
  data,
  file,
  title,
  pow2max,
  negpow2max,
  negcol,
  poscol,
  legendtitle,
  legYes,
  eps
)
}
\arguments{
\item{data}{array with data to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{pow2max}{upper (positive) end of data range to plot (2^pow2max)}

\item{negpow2max}{lower (negative) end of data range to plot (-2^negpow2max)}

\item{negcol}{color palette name from brewer.pal (character string) for negative values}

\item{poscol}{color palette name from brewer.pal (character string) for positive values}

\item{legendtitle}{character string legend title}

\item{legYes}{show legend (boolean)}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Creates a PNG/eps with a plot of a global LPJmL array
   Data is plotted in range: c(-2^negpow2max,2^pow2max)
   positive and negative values are colored accoding to chosen palettes,
   0-range is white,
}
\examples{
\dontrun{
plotGlobal_pos_neg(data=EFR_deficits2006,file=paste("~/","EFRdeficits_06.png",sep=""),
            title = "",negpow2max=10, pow2max=15,negcol="GnBu",poscol="Reds"
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)
}
}
