% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGlobalMan}
\alias{plotGlobalMan}
\title{Plot global LPJmL data with manual breaks and colorramp}
\usage{
plotGlobalMan(
  data,
  file,
  title,
  brks,
  palette = "YlGnBu",
  legendtitle,
  legYes,
  eps
)
}
\arguments{
\item{data}{LPJmL specific array c(67420) to be plotted}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{brks}{array with breaks to be used}

\item{palette}{color palette string (defaults to "YlGnBu")}

\item{legendtitle}{character string legend title}

\item{legYes}{plot legend (boolean)}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Creates a PNG/eps plot of a global LPJmL array dim=c(67420) with manual
breaks and a custom colorramp (if manual colorramp is not provided, YlGnBu is used)
}
\examples{
plotMonth(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep=""),
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)

}
