% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGlobalWflex}
\alias{plotGlobalWflex}
\title{Plot global LPJmL grid with flexibility to zoom into a region,}
\usage{
plotGlobalWflex(
  data,
  file,
  title,
  man = F,
  mbrk = -1,
  mpalette = -1,
  pow2max,
  pow2min,
  colPos = "Blues",
  colNeg = "Reds",
  legendtitle,
  legYes,
  onlyPos = F,
  eps,
  ires = 12,
  lon,
  lat,
  map = T,
  ext = c(-180, 180, -60, 90),
  logscale = T
)
}
\arguments{
\item{data}{array with month index to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}

\item{legYes}{plot legend (boolean)}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Creates a PNG/eps with a plot
}
\examples{
plotMonth(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep=""),
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)

}
