% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGlobalWlin}
\alias{plotGlobalWlin}
\title{Plot global LPJmL array}
\usage{
plotGlobalWlin(
  data,
  file,
  title,
  max,
  min,
  colPos = "Blues",
  colNeg = "Reds",
  colrev = F,
  legendtitle,
  legYes,
  onlyPos = F,
  eps
)
}
\arguments{
\item{data}{array with data to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{max}{upper (positive) end of data range to plot}

\item{min}{smallest positive number to be distinguished from 0}

\item{colPos}{color palette for the positives}

\item{colNeg}{color palette for the negatives}

\item{colrev}{reverse positive color palette if onlyPos (boolean)}

\item{legendtitle}{character string legend title}

\item{legYes}{show legend (boolean)}

\item{onlyPos}{show only positive (boolean)}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Creates a PNG/eps with a plot of a global LPJmL array
   Data is plotted linearly in range: c(-max,min,0,min,max)
   colors for pos and neg values can be given, default is Blues for the positive
   and Reds for the negative numbers
   0-range (from -min to min) is white.
   The negatives can be omitted by setting onlyPos=T, in case there are only pos values.
}
\examples{
plotGlobalWlin(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep="")max=15,min=0,
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)

}
