% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGlobalWtoscreen}
\alias{plotGlobalWtoscreen}
\title{Plot global LPJmL array inside RStudio}
\usage{
plotGlobalWtoscreen(data, title, pow2max, pow2min, legendtitle, legYes)
}
\arguments{
\item{data}{array with data to plot in LPJmL specific array c(67420)}

\item{title}{character string title for plot}

\item{pow2max}{upper (positive) end of data range to plot (2^pow2max)}

\item{pow2min}{smallest positive number to be distinguished from 0 (2^-pow2min)}

\item{legendtitle}{character string legend title}

\item{legYes}{show legend (boolean)}
}
\value{
None
}
\description{
Plot of a global LPJmL array inside RStudio
   Data is plotted in range: c(-2^pow2max,-2^-pow2min,0,2^-pow2min,2^pow2max)
   where the positive values are colored green to blue,
   0-range is white,
   and the negative ones red to yellow
}
\examples{
plotGlobalWtoscreen(data=irrigation2006,title = paste("irrigation amount 2006 in mm/yr",sep=""),
                    pow2max=15,pow2min=0,"legendtitle",legYes=TRUE)

}
