% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLUshares}
\alias{plotLUshares}
\title{Plot global LPJmL CFT array}
\usage{
plotLUshares(file, lushares, title, legendtitle, eps)
}
\arguments{
\item{file}{character string for location/file to save plot to}

\item{lushares}{array with cft band data to plot in LPJmL specific array c(67420)}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Plot of one band of a global LPJmL cft array
            classification into 10% classes.
}
\examples{
\dontrun{
plotLUshares(file="~/cftfracs2005_maize.png",lushares=cftfracs2005[,3],
             title = "cft fracs for maize",legendtitle="",eps=FALSE)
}

}
