% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLUshares2}
\alias{plotLUshares2}
\title{Plot global LPJmL CFT array}
\usage{
plotLUshares2(file, lushares, title, col, legendtitle, eps)
}
\arguments{
\item{file}{character string for location/file to save plot to}

\item{lushares}{array with cft band data to plot in LPJmL specific array c(67420)}

\item{title}{character string title for plot}

\item{col}{color palette name from brewer.pal (character string)}

\item{legendtitle}{character string legend title}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Plot of one band of a global LPJmL cft array
            classification into 10% classes, except for first class: 0-1%, 1-10%,
            colors according to chosen palette
}
\examples{
\dontrun{
plotLUshares2(file="~/cftfracs2005_maize.png",lushares=cftfracs2005[,3],
             title = "cft fracs for maize",col="Reds",legendtitle="",eps=FALSE)
}

}
