% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLUsharesToScreen}
\alias{plotLUsharesToScreen}
\title{Plot global LPJmL CFT array inside RStudio}
\usage{
plotLUsharesToScreen(lushares, title, legendtitle)
}
\arguments{
\item{lushares}{array with cft band data to plot in LPJmL specific array c(67420)}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}
}
\value{
None
}
\description{
Plot of one band of a global LPJmL cft array inside RStudio
           with classification into 10% classes.
}
\examples{
\dontrun{
plotLUsharesToScreen(lushares=cftfracs2005[,3],title = "cft fracs for maize",legendtitle="")
}

}
