% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMonth}
\alias{plotMonth}
\title{Plot global month indices on LPJmL grid, for example for months with highest precipitation}
\usage{
plotMonth(data, file, title, legendtitle, legYes, eps)
}
\arguments{
\item{data}{array with month index to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{legendtitle}{character string legend title}

\item{legYes}{plot legend (boolean)}

\item{eps}{write eps file instead of PNG (boolean)}
}
\value{
None
}
\description{
Creates a PNG/eps with a plot of a month index from 1:12
}
\examples{
plotMonth(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep=""),
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE)

}
