% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRegionalW}
\alias{plotRegionalW}
\title{Plot a zoomed in part from a LPJmL array}
\usage{
plotRegionalW(
  data,
  file,
  title,
  pow2max,
  pow2min,
  ext,
  colPos = "Blues",
  colNeg = "Reds",
  legendtitle,
  legYes,
  onlyPos = F,
  eps,
  ires = 12,
  lon,
  lat,
  map = T
)
}
\arguments{
\item{data}{array with data to plot in LPJmL specific array c(67420)}

\item{file}{character string for location/file to save plot to}

\item{title}{character string title for plot}

\item{pow2max}{upper (positive) end of data range to plot (2^pow2max)}

\item{pow2min}{smallest positive number to be distinguished from 0 (2^-pow2min)}

\item{ext}{extent to be plotted c(xmin,xmax,ymin,ymax)}

\item{colPos}{color palette for the positives}

\item{colNeg}{color palette for the negatives}

\item{legendtitle}{character string legend title}

\item{legYes}{show legend (boolean)}

\item{onlyPos}{show only positive (boolean)}

\item{eps}{write eps file instead of PNG (boolean)}

\item{ires}{resolution factor (2 for 30min, 12 for 5min)}

\item{lon}{longitude array}

\item{lat}{latitude array}

\item{map}{boolean whether to add country borders to the plot (defaults to T)}
}
\value{
None
}
\description{
Creates a PNG/eps with a plot of a global LPJmL array
   Data is plotted in range: c(-2^pow2max,-2^-pow2min,0,2^-pow2min,2^pow2max)
   colors for pos and neg values can be given, default is Blues for the positive
   and Reds for the negative numbers
   0-range (from 2^-pow2min to 2^pow2min) is white.
   The negatives can be omitted by setting onlyPos=T, in case there are only pos values.
}
\examples{
plotRegionalW(data=irrigation2006,file=paste("~/","mwateramount_2005_06.png",sep=""),
            title = paste("irrigation amount 2006 in mm/yr",sep=""),pow2max=15,pow2min=0,ext=c(120,160,-10,10),
            legendtitle="legendtitle",legYes=TRUE,eps=FALSE,ires=2,lon=lon,lat=lat,map=T)

}
