% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdf.R
\name{read_cdf}
\alias{read_cdf}
\title{Reads netcdf and returns it as array}
\usage{
read_cdf(filename, nc_header, subset = list(), silent = TRUE)
}
\arguments{
\item{filename}{netcdf file name}

\item{nc_header}{header data, read in from either meta file or data in
netcdf file}

\item{subset}{list object defining which subset of the data to be read}

\item{silent}{suppress warnings. Default: TRUE}
}
\value{
array with netcdf's data, dim=c(nlon,nlat,bands,steps (months/days),years)
}
\description{
Reads an arbitrary netcdf and returns the values at the lon/lat locations..
}
