% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cdf.R
\name{read_cdf_meta}
\alias{read_cdf_meta}
\title{Reads netcdf and returns it header info}
\usage{
read_cdf_meta(filename, variable_name, silent)
}
\arguments{
\item{filename}{netcdf file name}

\item{variable_name}{optional variable to be read, in case automatic
detection does not work as intended or several variables are stored within
the file}

\item{silent}{suppress warnings. Default: TRUE}
}
\value{
header data
}
\description{
Reads an arbitrary netcdf + returns the header values to read as lpjml object
}
