% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grid.R
\name{read_grid}
\alias{read_grid}
\title{Read LPJmL input and output grid files}
\usage{
read_grid(...)
}
\arguments{
\item{...}{See \link{read_io} for further arguments.}
}
\value{
An \link{LPJmLGridData} object.
}
\description{
Generic function to read LPJmL input & output files in different
formats. Depending on the format, arguments can be automatically detected
or have to be passed as individual arguments.
}
\details{
See \link{read_io} for more details.
}
\examples{
\dontrun{
my_grid <- read_io("grid.bin.json")

}
}
