% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_header}
\alias{read_header}
\title{Read header (any version) from LPJmL file.}
\usage{
read_header(filename, force_version = NULL)
}
\arguments{
\item{filename}{Filename to read header from}

\item{force_version}{Manually set clm version (default value 'NULL' means
that version is determined automatically from header; set only if version
number in file is incorrect)}
}
\value{
The function returns a list with 3 components:
\itemize{
\item name: header name, e.g. "LPJGRID"; describes the type of data in file
\item header: vector of header values ('version', 'order', 'firstyear',
'nyear', 'firstcell', 'ncell', 'nbands', 'cellsize_lon', 'scalar',
'cellsize_lat', 'datatype') describing the file structure; if header
version is <3, partially filled with default values
\item endian: endianness of file (little or big)
}
}
\description{
Reads a header from an LPJmL clm file. CLM is the default
format used for LPJmL input files and can also be used for output files.
}
\examples{
\dontrun{
header <- read_header("filename.clm")
}

}
\seealso{
\code{\link[=new_header]{new_header()}} for a more detailed description of the LPJmL header format,
\code{\link[=write_header]{write_header()}} for writing headers to files.
}
