% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_io.R
\name{read_input}
\alias{read_input}
\title{Read LPJmL input files}
\usage{
read_input(
  file_name,
  file_type = "meta",
  band_names = NULL,
  nstep = NULL,
  timestep = NULL,
  subset_list = list(),
  dim_order = c("cell", "time", "band"),
  version = NULL,
  order = NULL,
  firstyear = NULL,
  nyear = NULL,
  firstcell = NULL,
  ncell = NULL,
  nbands = NULL,
  cellsize_lon = NULL,
  scalar = NULL,
  cellsize_lat = NULL,
  datatype = NULL,
  endian = NULL,
  variable = NULL,
  descr = NULL,
  unit = NULL,
  name = NULL,
  silent = FALSE
)
}
\description{
Alias of \code{read_io}
}
\seealso{
\link{read_io}
}
