% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCFTinput.R
\name{readCFTinput}
\alias{readCFTinput}
\title{Read cft LPJmL input}
\usage{
readCFTinput(
  inFile,
  startyear,
  stopyear,
  bands,
  size,
  dtype,
  headersize = 43,
  getyearstart,
  getyearstop,
  ncells = 67420
)
}
\arguments{
\item{inFile}{character string containing the file to read the data from}

\item{startyear}{absolute startyear of output}

\item{stopyear}{absolute stopyear of output}

\item{bands}{number of cft-bands (32 for standard LPJmL input, 64 for standard input)}

\item{size}{size of each cell's data (2 or 4 bytes)}

\item{headersize}{size of header data (defaults to 43 bytes)}

\item{getyearstart}{start of range to return}

\item{getyearstop}{end of range to return}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}
}
\value{
array of cft-fractions c(bands,ncells,nyears) for each of the 67420 gridcells over requested range of years
}
\description{
Returns a range of years from a cft LPJmL input of dimension c(ncells,bands,nyears).
}
\examples{
\dontrun{
readCFTinput(inFile="cftinput.bin",startyear=1901,stopyear=2015,bands=64,size=4,headersize=43,
        getyearstart=2005,getyearstop=2005)
}

}
