% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCFToutput.R
\name{readCFToutput}
\alias{readCFToutput}
\title{Read cft LPJmL output}
\usage{
readCFToutput(
  inFile,
  startyear,
  stopyear,
  bands,
  size,
  headersize,
  getyearstart,
  getyearstop,
  ncells = 67420
)
}
\arguments{
\item{inFile}{character string containing the file to read the data from}

\item{startyear}{absolute startyear of output}

\item{stopyear}{absolute stopyear of output}

\item{bands}{number of bands (32 for standard LPJmL output, 64 for standard input)}

\item{size}{size of each cell's data (2 or 4 bytes)}

\item{headersize}{size of header data (defaults to 0 bytes)}

\item{getyearstart}{start of range to return}

\item{getyearstop}{end of range to return}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}
}
\value{
array of data c(ncells,bands,nyears) for each of the 67420 gridcells over requested range of years or c(ncells,bands) for one year
}
\description{
Returns a range of years from a cft LPJmL output of dimension c(ncells,bands,nyears).
}
\examples{
\dontrun{
readCFToutput(inFile="cftfrac.bin",startyear=1861,stopyear=2005,bands=32,size=4,headersize=0,
        getyearstart=2005,getyearstop=2005,ncells=67420)
}

}
