% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{Read a LPJmL config JSON file}
\usage{
readConfig(filename)
}
\arguments{
\item{filename}{character string representing path
(if differs from current working directory) and filename}
}
\value{
nested list object representing the structure of \code{config.json}
}
\description{
Reads a JSON file (not \code{lpjml.js}!, but its precompiled version
using \code{parseConfig} and turns it into a nested list object.
}
\examples{
\dontrun{
 config <- readConfig(filename = "config.json")
 
 config[["version"]]
 # [1] "5.3"

 config[["pftpar"]][[1]][["name"]]
 # [1] "tropical broadleaved evergreen tree"

 config[["input"]][["coord"]][["name"]]
 # [1] "input_VERSION2/grid.bin"
}
}
