% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readheader}
\alias{readheader}
\title{Reads a header from an LPJ input file}
\usage{
readheader(filename, force_version = NULL, endian = .Platform$endian)
}
\arguments{
\item{filename}{filename to read header from}

\item{force_version}{force clm version (default NULL - no)}

\item{endian}{endianness (default copied from local platform -- .Platform$endian)}
}
\value{
list-object with header content
}
\description{
tries to determine header version unless force_version is provided
return value: list with 3 components:                              ##
- header name, e.g. LPJGRID                                        ##
- header values (11 in total), if header version is <3, partially  ##
  filled with default values                                       ##
- endian of file (little or big)
}
\examples{
\dontrun{
readheader(filename, force_version=NULL)
}

}
