% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readMonthly}
\alias{readMonthly}
\title{Read monthly LPJmL output file}
\usage{
readMonthly(
  inFile,
  startyear,
  stopyear,
  size,
  headersize = 0,
  getyearstart,
  getyearstop,
  ncells = 67420
)
}
\arguments{
\item{inFile}{character string containing the file to read the data from}

\item{startyear}{absolute startyear of output}

\item{stopyear}{absolute stopyear of output}

\item{size}{size of each cell's data (2 or 4 bytes)}

\item{headersize}{size of header data (defaults to 0 bytes)}

\item{getyearstart}{start of range to return}

\item{getyearstop}{end of range to return}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}
}
\value{
array of monthly data c(ncells,12,nyears) for each of the 67420 gridcells over requested range of years
}
\description{
Returns a range of years from a monthly LPJmL output of dimension c(ncells,12,nyears)
}
\examples{
\dontrun{
readMonthly(inFile="mwateramount.bin",startyear=1861,stopyear=2005,size=4,headersize=0,
            getyearstart=1984,getyearstop=2005)
}

}
