% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readYearly}
\alias{readYearly}
\title{Read yearly LPJmL output file}
\usage{
readYearly(
  inFile,
  startyear,
  stopyear,
  size,
  headersize = 0,
  getyearstart,
  getyearstop,
  ncells = 67420
)
}
\arguments{
\item{inFile}{character string containing the file to read the data from}

\item{startyear}{absolute startyear of output}

\item{stopyear}{absolute stopyear of output}

\item{size}{size of each cell's data (2 or 4 bytes)}

\item{headersize}{size of header data (defaults to 0 bytes)}

\item{getyearstart}{start of range to return}

\item{getyearstop}{end of range to return}

\item{ncells}{number of lpj cells (67420 for 30min res, 2298847 for 5min res)}
}
\value{
array of data c(ncells,nyears) for each of the 67420 gridcells over requested range of years
}
\description{
Returns a range of years from a LPJmL output of dimension c(ncells,nyears)
}
\examples{
\dontrun{
readYearly(inFile="hdates.bin",startyear=1901,stopyear=2005,size=2,headersize=0,
           getyearstart=1984,getyearstop=2005)
}

}
