% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{removeHighFlows}
\alias{removeHighFlows}
\title{Remove highflows from a discharge input}
\usage{
removeHighFlows(ddischarge, threshold = 0.9, period = "all")
}
\arguments{
\item{ddischarge}{daily discharge input}

\item{threshold}{all highflows above this quantile (0.9 for Q90) will be removed}

\item{period}{calculate the highflows based on all days from all years ("all"), separately for each year ("year"), or just from the days which have positive discharge ("posDischarge")}
}
\value{
modified discharge array with the highflows removed
}
\description{
Remove highflows (above the given quantile) from a discharge input and return the modified discharge
}
\examples{
\dontrun{
removeHighFlows(ddischarge=daily_discharge_preindustrial,threshold=0.8,period="year")
}

}
