% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_LpjmlData.R
\name{reshape_LpjmlData}
\alias{reshape_LpjmlData}
\title{Reshape a LpjmlData object into an array with rearranged dimensions}
\usage{
reshape_LpjmlData(
  data_grid = NULL,
  meta_grid = NULL,
  data_lpjml = NULL,
  meta_lpjml = NULL,
  output_shape = "array_lon-lat-time"
)
}
\arguments{
\item{data_grid}{An LPJmL grid output in the form of LpjmLData$data}

\item{meta_grid}{An LPJmL grid output in the form of LpjmLData$meta_data}

\item{data_lpjml}{An LPJmL output in the form of LpjmLData$data
(note: one band only allowed!)}

\item{meta_lpjml}{An LPJmL output in the form of LpjmLData$meta_data
(note: one band only allowed!)}

\item{output_shape}{The shape of the output array. Options are:
"array_lon-lat-time" \link{lon, lat, time}; (TODO: other shapes to be implemeted)}
}
\value{

}
\description{
Function to read LpjmLData object and rearrange its dimensions
into a different structure. LpjmLData dimensions are \link{cell, time, band}. With
this function, you can convert that into an array with dimensions
\link{lon, lat, time} (option output_shape = "array_lon-lat-time").
}
\seealso{

}
