% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLPJmL.R
\name{runLPJmL}
\alias{runLPJmL}
\title{Run LPJmL}
\usage{
runLPJmL(sim_name_s, model_path, output_path = NULL)
}
\arguments{
\item{sim_name_s}{character vector providing sim_name or sim_names if the
runs are dependent from each other (e.g. spin-up and transient run)}

\item{model_path}{character string providing the path to LPJmL
(equal to LPJROOT)}

\item{output_path}{character string - if output_path differs from
model_path - path where an output, a restart and a configuration
folder are created}
}
\value{
"Done."
}
\description{
LPJmL is run using a config*.json written by \link[lpjmlKit]{writeConfig}.
\code{runLPJmL} can only perform single or subsequent model runs. For multiple
runs (via slurm) you may consider to use \link[lpjmlKit]{submitLPJmL} or
loop over an R-script containing \code{runLPJmL}.
}
