% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{showRoute}
\alias{showRoute}
\title{Show the full drainage route for cell ind}
\usage{
showRoute(ind, routingTable)
}
\arguments{
\item{ind}{cell index to look at}

\item{routingTable}{routing table obtained from reading in drainage network and adding 1, listing for each cell the next cell discharge is routed to and for sinks -0}
}
\value{
list of downstream cells
}
\description{
Show the full downstream drainage route for cell ind until final drainage (ocean or inland sink)
}
\examples{
\dontrun{
showRoute(ind=20000,routingTable=routing)
}

}
