% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_names.R
\name{split_time_names}
\alias{split_time_names}
\title{Split date strings into years, months and days}
\usage{
split_time_names(time_names)
}
\arguments{
\item{time_names}{Character vector with one or several date strings in the
form YYYY-MM-DD.}
}
\value{
List containing unique \code{year}, \code{month} and \code{day} values included in
\code{time_names}.
}
\description{
Splits one or several date strings into a list of unique days, months and
years.
}
\examples{
\dontrun{
 time_names <- split_time_names(c("2024-11-25", "2024-11-26"))
 time_names
 # $year
 # [1] "2024"
 # $month
 # [1] "11"
 # $day
 # [1] "25" "26"
}

}
