% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_subset.R
\name{subset}
\alias{subset}
\title{Subset LPJmLData object}
\usage{
subset(x, ...)
}
\arguments{
\item{...}{dimensions of each}
}
\value{
LPJmLData object
}
\description{
Use dimension names of LPJmLData array directly to subset each by simply
using subset vectors.
}
