% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_array.R
\name{subset_array}
\alias{subset_array}
\alias{asub}
\alias{replace_array}
\title{Subset a named array}
\usage{
subset_array(x, subset_list = NULL, drop = TRUE)
asub(x, subset_list = NULL, drop = TRUE)
replace_array(x, subset_list, y)
}
\arguments{
\item{x}{array with named dimensions}

\item{subset_list}{names list with names being names of dimnames to be
subsetted. list values can either be numeric (leads to subsetting by indices)
or character strings (leads to subsetting by dimnames). Default is \code{list()}}

\item{drop}{logical. If TRUE (default), dimensions are dropped when dimension
has length == 1, else dimension is kept.}

\item{y}{array/vector of the same dimension as referred to by the subset_list}
}
\value{
array or vector (if \code{drop=TRUE} and one only dimension left)
}
\description{
Subset an array with supplied dimnames and - if defined - replace it.
}
\examples{
my_array <- array(1,
                  dim=c(cell=67, month=12, band=3),
                  dimnames=list(cell=0:66,
                                month=1:12,
                                band=c("band1", "band2", "band3")))
my_subset <- subset_array(my_array,
                          subset_list = list(bands=c("band1", "band3")))
dimnames(my_subset)[3]
# $ band
#   [1] "band1"
#   [2] "band3"

# replace subset
my_replacement <- replace_array(my_subset,
                                subset_list = list(bands=c("band1")),
                                0)
}
