% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{summary.LPJmLData}
\alias{summary.LPJmLData}
\title{Function to get the summary of the array of a \code{LPJmLData} object.
See also \link[base]{summary}}
\usage{
\method{summary}{LPJmLData}(object, ...)
}
\arguments{
\item{object}{\link{LPJmLData} object}

\item{...}{Further arguments:
\code{dimension} for which a summary is printed for every element
(in style of matrix summary). Default is \verb{dimension = band"}. Choose from
available dimensions like \code{"time"} or \code{"cell"}.
\code{subset} list of array dimension(s) as name/key and corresponding subset
vector as value, e.g. \verb{list(cell = c(27411:27415)}, more information at
\code{\link{subset.LPJmLData}}.
\code{cutoff} (logical) If \code{TRUE} summary for dimension elements > 16 are
cut off.
Or arguments to be passed on to \link[base]{summary}}
}
\description{
Function to get the summary of the array of a \code{LPJmLData} object.
See also \link[base]{summary}
}
