% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{upstreamCells}
\alias{upstreamCells}
\title{Return next direct upstream cells for cell ind}
\usage{
upstreamCells(ind, routingTable)
}
\arguments{
\item{ind}{cell index to look at}

\item{routingTable}{routing table obtained from reading in drainage network and adding 1, listing for each cell the next cell discharge is routed to and for sinks -0}
}
\value{
list of direct (neighboring) upstream cells
}
\description{
Return a list of the next direct upstream cells for cell ind (draining into ind)
}
\examples{
\dontrun{
upstreamCells(ind=20000,routingTable=routing)
}

}
