% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_config.R
\name{view_config}
\alias{view_config}
\title{View a LPJmL Configuration file}
\usage{
view_config(config_file)
}
\arguments{
\item{config_file}{character string providing the file name and the path to
the config file}
}
\value{
opens a new \code{View} tab
}
\description{
Open a tab and show all configuration file content in a tree view. Branches
can be folded and unfolded for a better overview or a detailed view.
}
