% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCFTinput.R
\name{writeCFTinput}
\alias{writeCFTinput}
\title{Write cft LPJmL input file}
\usage{
writeCFTinput(filename, input_array, firstyear, lastyear, bands)
}
\arguments{
\item{filename}{character string containing the name of the file to write the cft input in}

\item{input_array}{array of dimension c(ncells,bands,nyears)}

\item{firstyear}{first year in array}

\item{lastyear}{last year in array}

\item{bands}{number of cft-bands (32 or 64 for standard input)}
}
\value{
boolean containing TRUE (the sums of all cft fracs per grid cell are always equal or below 1000) or FALSE (the sums of all cft fracs per grid cell are not always equal or below 1000)
}
\description{
Writes a binary CLM(2) LPJmL input file with header from an array of dimension c(ncells,bands,nyears)
and checks whether the sums of all cft fracs per grid cell are always <= 1000.
}
\examples{
\dontrun{
writeCFTinput(filename="cftfrac_rcp26_2006-2100_64bands.clm", input_array = CFT_2006_2100, firstyear=2006,lastyear=2100,bands=64)
}

}
