% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeheader.R
\name{writeheader}
\alias{writeheader}
\title{write a header to a file}
\usage{
writeheader(filename, header, overwrite = FALSE)
}
\arguments{
\item{filename}{filename to write header into}

\item{header}{header list}

\item{overwrite}{overwrites output file if exists (default FALSE)}
}
\value{
None
}
\description{
expects a list following the structure returned by readheader() or
new_header()
will fail if output file exists, unless overwrite set to TRUE
}
\examples{
\dontrun{
writeheader(filename="filename.clm", header=header, overwrite=FALSE)
}

}
