% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeWUinput}
\alias{writeWUinput}
\title{Write wateruse LPJmL input file}
\usage{
writeWUinput(
  filename,
  input_array,
  fromyear,
  toyear,
  out_bands,
  out_name,
  out_size,
  out_scaling = 1
)
}
\arguments{
\item{filename}{character string containing the name of the file to write the cft input in}

\item{input_array}{array of dimension c(ncells,bands,nyears) or c(ncells,nyears)}

\item{out_name}{LPJmL variable to write out e.g. "LPJLUSE" oder "LPJWUSE"}

\item{out_size}{size of integer to be written out (2/4)}

\item{out_scaling}{optional scaling factor to be applied, defaults to 1}

\item{firstyear}{first year in array}

\item{lastyear}{last year in array}

\item{bands}{number of cft-bands (32 or 64 for standard input)}
}
\description{
Writes a binary CLM(2) LPJmL input file with header from an array of dimension c(ncells,bands,nyears)
}
\examples{
\dontrun{
writeWUinput(filename="wateruse_wd_2band_2006_2050.clm", input_array = waterinput, firstyear=2006, lastyear=2050, bands=2, out_name="LPJWUSE", out_size=4, out_scaling=1)
}

}
