% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData.R
\name{add_grid}
\alias{add_grid}
\title{Add grid to LPJmLData object}
\usage{
add_grid(x, ...)
}
\arguments{
\item{x}{\link{LPJmLData} object}

\item{...}{arguments passed to \code{\link{read_io}} if no grid file and or meta
file in corresponding output directory available.}
}
\value{
\code{\link{LPJmLData}} object in selected format
}
\description{
Function to add a grid to a \code{\link{LPJmLData}}. The function acts
as a \code{\link{read_io}} function for the grid file and adds it as an
\code{LPJmLData} object itself to the the main object as an attribute \verb{$grid}
}
\details{
\strong{Important:}
\itemize{
\item If \code{"file_type" == "raw"} and data should be recognized as a
grid, prescribe \code{variable = "grid"}!
\item Do not use \link{read_io} argument \code{subset} here!
}
}
\examples{
\dontrun{

# read in vegetation carbon data with meta file
vegc <- read_io(filename = "./vegc.bin.json")

# add grid as attribute (via meta file in output directory)
vegc_with_grid <- add_grid(vegc)

}

}
