% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData.R
\name{add_grid}
\alias{add_grid}
\title{Add grid to LPJmLData object}
\usage{
add_grid(x, ...)
}
\arguments{
\item{to}{character vector defining space and/or time format into which
corresponding data dimensions should be transformed. Choose from space
formats \code{c("cell", "lon_lat")} and time formats \code{c("time","year_month_day")}.}
}
\value{
\link\href{LPJmLData}{lpjmlkit} object in selected format
}
\description{
Function to add a grid to an \link\href{LPJmLData}{lpjmlkit}. The function acts
as a \link\href{read_io}{lpjmlkit} function for the grid file and adds it as an
\code{LPJmLData} object itself to the the main object as an attribute \verb{$grid}
}
\examples{
\dontrun{

vegc <- read_io(filename = "./vegc.bin.json")

}

}
