% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{check_config}
\alias{check_config}
\title{Check LPJmL config JSON files using lpjcheck}
\usage{
check_config(x, model_path, output_path = NULL, pretty_print = TRUE)
}
\arguments{
\item{x}{job_details object returned by \link[lpjmlKit]{write_config} or
character vector providing the config file names
(hint: returns x as a list entry)}

\item{model_path}{character string providing the path to LPJmL
(equal to LPJROOT)}

\item{output_path}{character string - if output_path differs from
model_path - path where an output, a restart and a configuration
folder are created}

\item{pretty_print}{boolean - if TRUE print stdout/stderr else return.
Defaults to TRUE}
}
\value{
see \code{pretty_print}
}
\description{
Check if created LPJmL config JSON files (\link[lpjmlKit]{write_config}) are
valid.
}
