% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_type.R
\name{detect_type}
\alias{detect_type}
\title{Detect file type of a LPJmL file}
\usage{
detect_type(filename)
}
\arguments{
\item{filename}{Character string naming the file to check}
}
\value{
character vector of length 1 giving the file type:
\itemize{
\item "cdf" for a NetCDF file (classic or NetCDF4/HDF5 format)
\item "clm" for a binary LPJmL file with header
\item "meta" for a JSON meta file describing a LPJmL binary file
\item "raw" for a binary LPJmL file without header
\item "text" for any type of text-only file, e.g. \code{.txt} or \code{.csv}
}
}
\description{
This utility function tries to detect automatically if a
provided file is of "clm", "meta", or "raw" file type. NetCDFs and simple
text formats such as \code{.txt} or \code{.csv} are also detected.
}
