% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{dim.LPJmLData}
\alias{dim.LPJmLData}
\title{Dimensions of an LPJmLData data array
Function to get the dimensions of the data array of an LPJmLData object.}
\usage{
\method{dim}{LPJmLData}(x)
}
\arguments{
\item{x}{\link{LPJmLData} object}
}
\description{
Dimensions of an LPJmLData data array
Function to get the dimensions of the data array of an LPJmLData object.
}
