% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{dim.LPJmLData}
\alias{dim.LPJmLData}
\title{Get dimension names and lengths of LPJmLData's data array
Function to get the dimension names and lengths of the array of a LPJmLData
object}
\usage{
\method{dim}{LPJmLData}(x)
}
\arguments{
\item{x}{\link{LPJmLData} object}
}
\description{
Get dimension names and lengths of LPJmLData's data array
Function to get the dimension names and lengths of the array of a LPJmLData
object
}
