% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{dimnames.LPJmLData}
\alias{dimnames.LPJmLData}
\title{Dimnames of an LPJmLData data array}
\usage{
\method{dimnames}{LPJmLData}(x)
}
\arguments{
\item{x}{\link{LPJmLData} object}
}
\value{
A list of the same length as dim(x). Components are character vectors
with positive length of the appropriate dimension of x.
}
\description{
Function to get the dimnames (list) of the data array of an LPJmLData object.
}
