% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData.R
\name{dimnames.LPJmLData}
\alias{dimnames.LPJmLData}
\title{Get dimensions of LPJmLData's data array
Function to get the dimensions (list) of the array of a
\code{LPJmLData} object.}
\usage{
\method{dimnames}{LPJmLData}(x, ...)
}
\arguments{
\item{...}{further arguments to be passed on to \link[base]{dimnames}}
}
\description{
Get dimensions of LPJmLData's data array
Function to get the dimensions (list) of the array of a
\code{LPJmLData} object.
}
