% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{dimnames.LPJmLData}
\alias{dimnames.LPJmLData}
\title{Dimnames of an LPJmLData data array}
\usage{
\method{dimnames}{LPJmLData}(x)
}
\arguments{
\item{x}{\link{LPJmLData} object}
}
\description{
Function to get the dimnames (list) of the data array of an LPJmLData object.
}
