% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{headersize}
\alias{headersize}
\title{Computes header size from header list}
\usage{
headersize(header)
}
\arguments{
\item{header}{header list}
}
\value{
byte size of header
}
\description{
this function can be used within seek() to skip over a header
(e.g. from readheader()) in a LPJmL input file
}
\examples{
\dontrun{
headersize(header=header)
}

}
