% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_stats.R
\name{length.LPJmLData}
\alias{length.LPJmLData}
\title{Length of an LPJmLData data array}
\usage{
\method{length}{LPJmLData}(x)
}
\arguments{
\item{x}{\link{LPJmLData} object}
}
\value{
A non-negative integer or numeric (which will be rounded down).
}
\description{
Function to get the length of the data array of an LPJmLData object.
}
