% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj_datatype.R
\name{lpj_datatype}
\alias{lpj_datatype}
\title{Data type and size}
\usage{
lpj_datatype(header)
}
\arguments{
\item{header}{Header list object as returned by read_header() or new_header()}
}
\value{
The function returns a list object with three components:
\itemize{
\item type: R data type; can be used with 'what' parameter of readBin()
\item size: size of data type; can be used with 'size' parameter of readBin()
\item signed: whether or not the data type is signed, only really relevant for integers with size 1 or 2; can be used with 'signed' parameter of readBin()
}
}
\description{
Provide information on the data type used in an LPJmL file
based on the 'datatype' attribute included in the file header.
}
\examples{
\dontrun{
# read file header
header <- read_header("filename.clm")
# open file for reading
fp <- file("filename.clm", "rb")
# skip over file header
seek(fp, header_size(header))
# read in file data
file_data <- readBin(fp, what=lpj_datatype(header)$type, size=lpj_datatype(header)$size, signed=lpj_datatype(header)$signed, n=header[["header"]]["ncell"]*header[["header"]]["nbands"]*header[["header"]]["nyear"], endian=header[["endian"]])
# close file
close(fp)
}

}
\seealso{
\code{\link[=read_header]{read_header()}} for reading headers from LPJmL files, \code{\link[=new_header]{new_header()}} for creating headers from scratch, \code{\link[=header_size]{header_size()}} for determining the size of file headers
}
