% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LpjmlData.R
\name{LpjmlData}
\alias{LpjmlData}
\title{LPJmL data class}
\value{
LpjmlData object
}
\description{
Handles LPJmL output and input data
}
\examples{
\dontrun{
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LpjmlData$new()}}
\item \href{#method-as_tibble}{\code{LpjmlData$as_tibble()}}
\item \href{#method-as_array}{\code{LpjmlData$as_array()}}
\item \href{#method-as_raster}{\code{LpjmlData$as_raster()}}
\item \href{#method-as_rast}{\code{LpjmlData$as_rast()}}
\item \href{#method-length}{\code{LpjmlData$length()}}
\item \href{#method-dim}{\code{LpjmlData$dim()}}
\item \href{#method-dimnames}{\code{LpjmlData$dimnames()}}
\item \href{#method-subset}{\code{LpjmlData$subset()}}
\item \href{#method-add_grid}{\code{LpjmlData$add_grid()}}
\item \href{#method-convert_time}{\code{LpjmlData$convert_time()}}
\item \href{#method-summary}{\code{LpjmlData$summary()}}
\item \href{#method-print}{\code{LpjmlData$print()}}
\item \href{#method-clone}{\code{LpjmlData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$new(data_array, meta_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_array}}{LPJmL data array}

\item{\code{meta_data}}{LpjmlMetaData Object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$as_tibble(subset_list = NULL, value_name = "value")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_array"></a>}}
\if{latex}{\out{\hypertarget{method-as_array}{}}}
\subsection{Method \code{as_array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$as_array(subset_list = NULL, drop = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_raster"></a>}}
\if{latex}{\out{\hypertarget{method-as_raster}{}}}
\subsection{Method \code{as_raster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$as_raster(subset_list = NULL, fix_extent = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_rast"></a>}}
\if{latex}{\out{\hypertarget{method-as_rast}{}}}
\subsection{Method \code{as_rast()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$as_rast(grid_file, as_layers = "band", subset_list = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\if{latex}{\out{\hypertarget{method-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\if{latex}{\out{\hypertarget{method-dim}{}}}
\subsection{Method \code{dim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$dim()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dimnames"></a>}}
\if{latex}{\out{\hypertarget{method-dimnames}{}}}
\subsection{Method \code{dimnames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$dimnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\if{latex}{\out{\hypertarget{method-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$subset(subset_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_grid"></a>}}
\if{latex}{\out{\hypertarget{method-add_grid}{}}}
\subsection{Method \code{add_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$add_grid(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_time"></a>}}
\if{latex}{\out{\hypertarget{method-convert_time}{}}}
\subsection{Method \code{convert_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$convert_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$summary(dimension = "band", subset_list = NULL, cutoff = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
