% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LpjmlData.R
\name{LpjmlData}
\alias{LpjmlData}
\title{LPJmL data class}
\value{
LpjmlData object
}
\description{
Handles LPJmL output and input data
}
\examples{
\dontrun{
}
}
\section{Super class}{
\code{\link[lpjmlKit:LpjmlMetaData]{lpjmlKit::LpjmlMetaData}} -> \code{LpjmlData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LpjmlData$new()}}
\item \href{#method-[}{\code{LpjmlData$[()}}
\item \href{#method-length}{\code{LpjmlData$length()}}
\item \href{#method-dim}{\code{LpjmlData$dim()}}
\item \href{#method-fields}{\code{LpjmlData$fields()}}
\item \href{#method-print}{\code{LpjmlData$print()}}
\item \href{#method-clone}{\code{LpjmlData$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="lpjmlKit" data-topic="LpjmlMetaData" data-id="header">}\href{../../lpjmlKit/html/LpjmlMetaData.html#method-header}{\code{lpjmlKit::LpjmlMetaData$header()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="lpjmlKit" data-topic="LpjmlMetaData" data-id="list">}\href{../../lpjmlKit/html/LpjmlMetaData.html#method-list}{\code{lpjmlKit::LpjmlMetaData$list()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$new(data_array, meta_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_array}}{LPJmL data array}

\item{\code{meta_data}}{LpjmlMetaData Object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-["></a>}}
\if{latex}{\out{\hypertarget{method-[}{}}}
\subsection{Method \code{[()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$[(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\if{latex}{\out{\hypertarget{method-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\if{latex}{\out{\hypertarget{method-dim}{}}}
\subsection{Method \code{dim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$dim()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fields"></a>}}
\if{latex}{\out{\hypertarget{method-fields}{}}}
\subsection{Method \code{fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
