% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData.R
\name{LPJmLData}
\alias{LPJmLData}
\title{LPJmL data class}
\description{
LPJmL data class

LPJmL data class
}
\details{
Handles LPJmL output and input data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{`LPJmLMetaData`}}{object to store corresponding meta data}

\item{\code{Data}}{\code{array} containing the underlying data}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{`LPJmLMetaData`}}{object to store corresponding meta data}

\item{\code{Data}}{\code{array} containing the underlying data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LPJmLData$new()}}
\item \href{#method-add_grid}{\code{LPJmLData$add_grid()}}
\item \href{#method-aggregate_array}{\code{LPJmLData$aggregate_array()}}
\item \href{#method-length}{\code{LPJmLData$length()}}
\item \href{#method-dim}{\code{LPJmLData$dim()}}
\item \href{#method-dimnames}{\code{LPJmLData$dimnames()}}
\item \href{#method-summary}{\code{LPJmLData$summary()}}
\item \href{#method-print}{\code{LPJmLData$print()}}
\item \href{#method-as_array}{\code{LPJmLData$as_array()}}
\item \href{#method-as_tibble}{\code{LPJmLData$as_tibble()}}
\item \href{#method-as_raster}{\code{LPJmLData$as_raster()}}
\item \href{#method-as_terra}{\code{LPJmLData$as_terra()}}
\item \href{#method-plot}{\code{LPJmLData$plot()}}
\item \href{#method-._plot_by_band}{\code{LPJmLData$._plot_by_band()}}
\item \href{#method-subset}{\code{LPJmLData$subset()}}
\item \href{#method-transform}{\code{LPJmLData$transform()}}
\item \href{#method-transform_grid}{\code{LPJmLData$transform_grid()}}
\item \href{#method-transform_space}{\code{LPJmLData$transform_space()}}
\item \href{#method-transform_time}{\code{LPJmLData$transform_time()}}
\item \href{#method-clone}{\code{LPJmLData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new LPJmLData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$new(data_array, meta_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_array}}{\code{array} with LPJmL data}

\item{\code{meta_data}}{\code{LPJmLMetaData} Object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{LPJmLData} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_grid"></a>}}
\if{latex}{\out{\hypertarget{method-add_grid}{}}}
\subsection{Method \code{add_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$add_grid(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aggregate_array"></a>}}
\if{latex}{\out{\hypertarget{method-aggregate_array}{}}}
\subsection{Method \code{aggregate_array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$aggregate_array(aggregate_list = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\if{latex}{\out{\hypertarget{method-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\if{latex}{\out{\hypertarget{method-dim}{}}}
\subsection{Method \code{dim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$dim()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dimnames"></a>}}
\if{latex}{\out{\hypertarget{method-dimnames}{}}}
\subsection{Method \code{dimnames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$dimnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$summary(dimension = "band", subset = NULL, cutoff = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_array"></a>}}
\if{latex}{\out{\hypertarget{method-as_array}{}}}
\subsection{Method \code{as_array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$as_array(subset = NULL, aggregate = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$as_tibble(subset = NULL, aggregate = NULL, value_name = "value", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_raster"></a>}}
\if{latex}{\out{\hypertarget{method-as_raster}{}}}
\subsection{Method \code{as_raster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$as_raster(subset = NULL, aggregate = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_terra"></a>}}
\if{latex}{\out{\hypertarget{method-as_terra}{}}}
\subsection{Method \code{as_terra()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$as_terra(grid_file, as_layers = "band", subset = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$plot(subset = NULL, aggregate = NULL, raster_extent = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-._plot_by_band"></a>}}
\if{latex}{\out{\hypertarget{method-._plot_by_band}{}}}
\subsection{Method \code{._plot_by_band()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$._plot_by_band(data, dots)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\if{latex}{\out{\hypertarget{method-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$subset(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$transform(to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform_grid"></a>}}
\if{latex}{\out{\hypertarget{method-transform_grid}{}}}
\subsection{Method \code{transform_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$transform_grid(to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform_space"></a>}}
\if{latex}{\out{\hypertarget{method-transform_space}{}}}
\subsection{Method \code{transform_space()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$transform_space(to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform_time"></a>}}
\if{latex}{\out{\hypertarget{method-transform_time}{}}}
\subsection{Method \code{transform_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$transform_time(to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
