% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjmlkit-package.R
\docType{package}
\name{lpjmlkit-package}
\alias{lpjmlkit}
\alias{lpjmlkit-package}
\title{lpjmlkit: Function kit for basic LPJmL handling}
\description{
Function kit for basic LPJmL handling, including simulations, corresponding configurations as well as input & output data.
}
