% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjmlkit-package.R
\docType{package}
\name{lpjmlkit-package}
\alias{lpjmlkit}
\alias{lpjmlkit-package}
\title{lpjmlkit: Function kit for basic LPJmL handling}
\description{
A collection of base functions to facilitate the work with the DGVM LPJmL hosted at the Potsdam Institute for Climate Impact Research. It provides functions for running LPJmL, as well as reading, processing and writing model-related data such as inputs and outputs or configuration files.
}
\author{
\strong{Maintainer}: Jannes Breier \email{jannesbr@pik-potsdam.de}

Authors:
\itemize{
  \item Christoph Mueller \email{cmueller@pik-potsdam.de}
  \item Fabian Stenzel \email{stenzel@pik-potsdam.de}
  \item Sara Minoli \email{minoli@pik-potsdam.de}
  \item Sebastian Ostberg \email{ostberg@pik-potsdam.de}
  \item Stephen Wirth \email{wirth@pik-potsdam.de}
}

}
