% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjmlkit-package.R
\docType{package}
\name{lpjmlkit-package}
\alias{lpjmlkit}
\alias{lpjmlkit-package}
\title{lpjmlkit: Toolkit for basic LPJmL handling}
\description{
A collection of base functions to facilitate the work with the DGVM LPJmL hosted at the Potsdam Institute for Climate Impact Research. It provides functions for performing LPJmL simulations, as well as reading, processing and writing model-related data such as inputs and outputs or configuration files.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/PIK-LPJmL/lpjmlkit/issues, https://gitlab.pik-potsdam.de/lpjml/lpjmlkit/issues}
}

}
\author{
\strong{Maintainer}: Jannes Breier \email{jannesbr@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9055-6904}{ORCID})

Authors:
\itemize{
  \item Sebastian Ostberg \email{ostberg@pik-potsdam.de} (\href{https://orcid.org/0000-0002-2368-7015}{ORCID})
  \item Stephen Wirth \email{wirth@pik-potsdam.de} (\href{https://orcid.org/0000-0002-2261-8771}{ORCID})
  \item Sara Minoli \email{minoli@pik-potsdam.de} (\href{https://orcid.org/0000-0001-7920-3107}{ORCID})
  \item Fabian Stenzel \email{stenzel@pik-potsdam.de} (\href{https://orcid.org/0000-0002-5109-0048}{ORCID})
  \item Christoph Mueller \email{cmueller@pik-potsdam.de} (\href{https://orcid.org/0000-0002-9491-3550}{ORCID})
}

}
