% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LpjmlMetaData.R
\name{LpjmlMetaData}
\alias{LpjmlMetaData}
\title{LPJmL meta output class}
\value{
LpjmlMetaData object
}
\description{
Handles metafile data for output data
}
\examples{
\dontrun{
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LpjmlMetaData$new()}}
\item \href{#method-header}{\code{LpjmlMetaData$header()}}
\item \href{#method-fields}{\code{LpjmlMetaData$fields()}}
\item \href{#method-list}{\code{LpjmlMetaData$list()}}
\item \href{#method-print}{\code{LpjmlMetaData$print()}}
\item \href{#method-clone}{\code{LpjmlMetaData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$new(meta_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meta_list}}{list (not nested) with meta data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-header"></a>}}
\if{latex}{\out{\hypertarget{method-header}{}}}
\subsection{Method \code{header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$header()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fields"></a>}}
\if{latex}{\out{\hypertarget{method-fields}{}}}
\subsection{Method \code{fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
