% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LpjmlMetaData.R
\name{LpjmlMetaData}
\alias{LpjmlMetaData}
\title{LPJmL meta output class}
\value{
LpjmlMetaData object
}
\description{
Handles metafile data for output data
}
\examples{
\dontrun{
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LpjmlMetaData$new()}}
\item \href{#method-update_subset}{\code{LpjmlMetaData$update_subset()}}
\item \href{#method-as_header}{\code{LpjmlMetaData$as_header()}}
\item \href{#method-as_list}{\code{LpjmlMetaData$as_list()}}
\item \href{#method-print}{\code{LpjmlMetaData$print()}}
\item \href{#method-clone}{\code{LpjmlMetaData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$new(x, subset_list = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{list (not nested) with meta data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_subset"></a>}}
\if{latex}{\out{\hypertarget{method-update_subset}{}}}
\subsection{Method \code{update_subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$update_subset(subset_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_header"></a>}}
\if{latex}{\out{\hypertarget{method-as_header}{}}}
\subsection{Method \code{as_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$as_header()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$print(spaces = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LpjmlMetaData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
