% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLMetaData.R
\name{LPJmLMetaData}
\alias{LPJmLMetaData}
\title{LPJmL meta output class}
\description{
LPJmL meta output class

LPJmL meta output class
}
\details{
Handles metafile data for output data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LPJmLMetaData$new()}}
\item \href{#method-._update_subset}{\code{LPJmLMetaData$._update_subset()}}
\item \href{#method-._init_grid}{\code{LPJmLMetaData$._init_grid()}}
\item \href{#method-._transform_time_format}{\code{LPJmLMetaData$._transform_time_format()}}
\item \href{#method-._transform_space_format}{\code{LPJmLMetaData$._transform_space_format()}}
\item \href{#method-print}{\code{LPJmLMetaData$print()}}
\item \href{#method-as_header}{\code{LPJmLMetaData$as_header()}}
\item \href{#method-as_list}{\code{LPJmLMetaData$as_list()}}
\item \href{#method-clone}{\code{LPJmLMetaData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new LPJmLMetaData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$new(
  x,
  subset = list(),
  additional_data = list(),
  data_dir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{list} (not nested) with meta data}

\item{\code{subset}}{\code{list} of array dimension(s) as name/key and
corresponding subset vector as value, e.g.
\verb{list(cell = c(27411:27415)}, more information at
\link[lpjmlkit]{subset}.}

\item{\code{additional_data}}{\code{list} of additional attributes to be set that
are not included in file header. These are
\verb{c"(band_names", "variable", "descr", "unit")}}

\item{\code{data_dir}}{Character string for data directory to "lazy load" grid}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{LPJmLMetaData} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-._update_subset"></a>}}
\if{latex}{\out{\hypertarget{method-._update_subset}{}}}
\subsection{Method \code{._update_subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$._update_subset(
  subset,
  time_dimnames = NULL,
  cell_dimnames = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-._init_grid"></a>}}
\if{latex}{\out{\hypertarget{method-._init_grid}{}}}
\subsection{Method \code{._init_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$._init_grid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-._transform_time_format"></a>}}
\if{latex}{\out{\hypertarget{method-._transform_time_format}{}}}
\subsection{Method \code{._transform_time_format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$._transform_time_format(time_format)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-._transform_space_format"></a>}}
\if{latex}{\out{\hypertarget{method-._transform_space_format}{}}}
\subsection{Method \code{._transform_space_format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$._transform_space_format(space_format)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$print(all = TRUE, spaces = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_header"></a>}}
\if{latex}{\out{\hypertarget{method-as_header}{}}}
\subsection{Method \code{as_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$as_header(silent = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LPJmLMetaData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
