% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lpjml.R
\name{make_lpjml}
\alias{make_lpjml}
\title{Make all (compile, link, etc.) LPJmL}
\usage{
make_lpjml(model_path = ".", make_fast = TRUE, make_clean = FALSE)
}
\arguments{
\item{model_path}{character string providing the path to LPJmL
(equal to LPJROOT)}

\item{make_fast}{logical - if TRUE \verb{make -j8 all} is executed.
Defaults to TRUE}

\item{make_clean}{logical - if TRUE execute make clean first.
Defaults to FALSE}
}
\value{
a list with process status, see \link[processx]{run}
}
\description{
Compiles LPJmL and create executables by internally executing make all.
}
