% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lpjml.R
\name{make_lpjml}
\alias{make_lpjml}
\title{Make all (compile, link, etc.) LPJmL}
\usage{
make_lpjml(model_path = ".", make_fast = TRUE, pretty_print = TRUE)
}
\arguments{
\item{model_path}{character string providing the path to LPJmL
(equal to LPJROOT)}

\item{make_fast}{boolean - if TRUE \verb{make -j8 all} is executed.
Defaults to TRUE}

\item{pretty_print}{boolean - if TRUE print stdout/stderr else return.
Defaults to TRUE}
}
\value{
see \code{pretty_print}
}
\description{
Compiles LPJmL and create executables by internally executing make all.
}
