% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lpjml.R
\name{make_lpjml}
\alias{make_lpjml}
\title{Compile LPJmL model}
\usage{
make_lpjml(
  model_path = ".",
  make_fast = TRUE,
  make_clean = FALSE,
  throw_error = TRUE,
  debug = FALSE
)
}
\arguments{
\item{model_path}{Character string providing the path to LPJmL
(equal to LPJROOT)}

\item{make_fast}{Logical - if TRUE \verb{make -j16 all} is executed which uses 16
processors for compilation in parallel mode. Defaults to TRUE}

\item{make_clean}{Logical - if TRUE execute make clean first to remove
previous installation. Defaults to FALSE}

\item{throw_error}{Logical - if \code{FALSE} does not throw an error if sub
process has non-zero exit status, hence if compilation fails in first
attempt. Defaults to TRUE}

\item{debug}{Logical - if \code{TRUE}LPJmL is compiled with flag \code{-debug}}
}
\value{
A list with process status, see \link[processx]{run}
}
\description{
Compiles the LPJmL source code and creates an executable by internally
executing make all.
}
\examples{
\dontrun{
model_path <- "./LPJmL_internal"
make_lpjml(model_path = model_path)
}

}
