% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Read a LPJmL config file (JSON)}
\usage{
read_config(filename)
}
\arguments{
\item{filename}{character string representing path
(if differs from current working directory) and filename}
}
\value{
nested list object representing the structure of \code{config_*.json}
}
\description{
Reads a config JSON file (not \code{lpjml.js}!, but its precompiled version
using \code{parse_config} and turns it into a nested list object.
}
\examples{
\dontrun{
 config <- read_config(filename = "config_spinup.json")

 config[["version"]]
 # [1] "5.3"

 config[["pftpar"]][[1]][["name"]]
 # [1] "tropical broadleaved evergreen tree"

 config[["input"]][["coord"]][["name"]]
 # [1] "input_VERSION2/grid.bin"
}
}
