% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meta.R
\name{read_meta}
\alias{read_meta}
\title{Read a LPJmL meta file}
\usage{
read_meta(filename)
}
\arguments{
\item{filename}{character string representing path
(if differs from current working directory) and filename}
}
\value{
LpjmlMetaData object
}
\description{
Reads a meta JSON file.
}
\examples{
\dontrun{
 meta <- read_meta(filename = "mpft_npp.bin.json")

 meta[[sim_name]]
 # [1] "LPJmL Run"

 meta[[firstcell]]
 # [1] 27410

 meta[[pft]][[1]]
 # [1] "tropical broadleaved evergreen tree"
}
}
