% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_header_item.R
\name{set_header_item}
\alias{set_header_item}
\title{Set information in an LPJmL file header}
\usage{
set_header_item(header, ...)
}
\arguments{
\item{header}{LPJmL file header as returned by \code{read_header()} or
\code{create_header()}.}

\item{...}{Header items to set. Can be one or several of 'name', 'version',
'order', 'firstyear', 'nyear', 'firstcell', 'ncell', 'nbands',
'cellsize_lon', 'scalar', 'cellsize_lat', 'datatype', 'nstep', 'timestep',
'endian'.}
}
\value{
Header 'header' where items included in parameters have been changed.
}
\description{
Convenience function to set information in a header object as
returned by \code{read_header()} or \code{create_header()}. You can set one or several
}
\seealso{
\itemize{
\item \code{\link[=create_header]{create_header()}} for creating headers from scratch and for a more
detailed description of the LPJmL header format.
\item \code{\link[=read_header]{read_header()}} for reading headers from files.
}
}
