% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_subset.R
\name{subset.LPJmLData}
\alias{subset.LPJmLData}
\title{Subset LPJmLData object}
\usage{
\method{subset}{LPJmLData}(x, ...)
}
\arguments{
\item{...}{Provide dimension to bet subset of LPJmLData objects underlying
data array in combination with indices vectors, e.g. \code{cell = c(27409, 27415)}
, or ``band = -c(14:16, 19:32) or subset using a "character" vector like
\code{band = c("rainfed rice","rainfed maize")}}
}
\value{
LPJmLData object
}
\description{
Use dimension names of LPJmLData array directly to subset each by simply
using subset vectors.
}
\examples{

}
