% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_subset.R
\name{subset.LPJmLData}
\alias{subset.LPJmLData}
\title{Subset LPJmLData object}
\usage{
\method{subset}{LPJmLData}(x, ...)
}
\arguments{
\item{x}{\link{LPJmLData} object}

\item{...}{Provide dimension to be used to subset \code{LPJmLData} objects
underlying data array in combination with indices vectors,
e.g. \code{cell = c(27411:27416)}, or \code{band = -c(14:16, 19:32)} or subset
using a "character" vector like \code{band = c("rainfed rice","rainfed maize")}.}
}
\value{
\code{\link{LPJmLData}} object with/without subset cell of dimension
}
\description{
Function to use dimension names of \code{\link{LPJmLData}} array directly
to subset each by simply using supplying vectors.
}
\examples{
\dontrun{

vegc <- read_io(filename = "./vegc.bin.json")

# subset cells by index
subset(vegc, cell = c(27409, 27415))
# [...]
# $data \%>\%
#   dimnames() \%>\%
#     .$cell  "27409" "27410" "27411" "27412" "27413" "27414"
#     .$time  "1901-12-31" "1902-12-31" "1903-12-31" "1904-12-31" ...
#     .$band  "1"
# [...]

#' # subset time by character vector
subset(vegc, cell = c("2001-12-31", "2002-12-31", "2003-12-31"))
# [...]
# $data \%>\%
#   dimnames() \%>\%
#     .$cell  "0" "1" "2" "3" ... "67419"
#     .$time  "2001-12-31" "2002-12-31" "2003-12-31"
#     .$band  "1"
# [...]
}

}
