% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJmLData_transform.R
\name{transform.LPJmLData}
\alias{transform.LPJmLData}
\title{Transform LPJmLData object}
\usage{
\method{transform}{LPJmLData}(x, ...)
}
\arguments{
\item{to}{character vector defining space and/or time format into which
corresponding data dimensions should be transformed. Choose from space
formats \code{c("cell", "lon_lat")} and time formats \code{c("time","year_month_day")}.}
}
\value{
LPJmLData object
}
\description{
...
}
\examples{

}
