% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHeader.R
\name{writeHeader}
\alias{writeHeader}
\title{Write header (any version) to LPJmL file}
\usage{
writeHeader(filename, header, overwrite = FALSE)
}
\arguments{
\item{filename}{filename to write header into}

\item{header}{The header to be written}

\item{overwrite}{Whether to overwrite an existing output file (default FALSE)}
}
\value{
None
}
\description{
Write an LPJmL clm header to a file. The header has to be
a list following the structure returned by readHeader() or newHeader().
The function will fail if the output file exists already unless overwrite is set to TRUE.
}
\examples{
\dontrun{
header <- readHeader(filename="old_filename.clm")
writeHeader(filename="new_filename.clm", header=header, overwrite=FALSE)
}

}
\seealso{
\code{\link[=newHeader]{newHeader()}} for creating headers from scratch and for a more detailed
description of the LPJmL header format, \code{\link[=readHeader]{readHeader()}} for reading headers from files.
}
